feat: 支持ts
