+++ .eslintrc.js
@@ -1,0 +1,20 @@
+module.exports = {
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint'],
+  parserOptions: {
+    sourceType: 'module'
+  },
+  rules: {
+    // @fixable 必须使用 === 或 !==，禁止使用 == 或 !=，与 null 比较时除外
+    eqeqeq: [
+      'error',
+      'always',
+      {
+        null: 'ignore'
+      }
+    ],
+    '@typescript-eslint/indent': ['error', 2],
+    // 类和接口的命名必须遵守帕斯卡命名法，比如 PersianCat
+    '@typescript-eslint/class-name-casing': 'error'
+  }
+};
