+++ package.json
@@ -8,6 +8,7 @@
   "typings": "typings/index.d.ts",
   "scripts": {
     "start": "node ./index.js",
+    "ts:build": "tsc",
     "wda": "sh script/test.sh",
     "wda-init": "cd WebDriverAgent && ./Scripts/bootstrap.sh",
     "wda-clone": "git clone --depth=1 https://github.com/flasco/WebDriverAgent",
@@ -16,29 +17,40 @@
     "screenshot": "node __test__/screenshot.test.js",
     "test": "jest --no-cache --forceExit --runInBand --detectOpenHandles"
   },
-  "lint-staged": {
-    "*.js": [
-      "eslint",
-      "git add"
-    ]
-  },
-  "dependencies": {
-    "@flasco/logger": "git+https://github.com/flasco/logger.git",
-    "colors-cli": "^1.0.21",
-    "date-fns": "^1.30.1",
-    "opencv4nodejs": "^5.0.0"
-  },
-  "devDependencies": {
-    "@flasco/wda-driver": "git+https://github.com/flasco/WDR.git",
-    "eslint": "^5.6.0",
-    "husky": "^1.0.0-rc.14",
-    "jest": "23.6.0",
-    "lint-staged": "^7.3.0"
-  },
   "husky": {
     "hooks": {
       "pre-commit": "lint-staged"
     }
+  },
+  "lint-staged": {
+    "linters": {
+      "*.{js,ts}": [
+        "eslint",
+        "git add"
+      ]
+    },
+    "ignore": [
+      "**/dist/**/*",
+      ".eslint*.*"
+    ]
+  },
+  "dependencies": {
+    "@flasco/logger": "git+https://github.com/flasco/logger.git#feature/support-ts",
+    "colors-cli": "^1.0.21",
+    "date-fns": "^1.30.1",
+    "opencv4nodejs": "^5.0.0",
+    "tslib": "^1.10.0"
+  },
+  "devDependencies": {
+    "@flasco/wda-driver": "git+https://github.com/flasco/WDR.git#feature/support-ts",
+    "@types/node": "^12.0.10",
+    "@typescript-eslint/eslint-plugin": "^1.11.0",
+    "@typescript-eslint/parser": "^1.11.0",
+    "eslint": "^6.0.1",
+    "husky": "^2.7.0",
+    "jest": "^24.8.0",
+    "lint-staged": "^8.2.1",
+    "typescript": "^3.5.2"
   },
   "jest": {
     "testEnvironment": "node",
