+++ src/index.ts
@@ -1,0 +1,5 @@
+import * as chainOperation from './utils/chainOperation';
+import { delay, base642Mat } from './utils';
+import Judge from './core/base-judge';
+
+export { delay, base642Mat, Judge, chainOperation };
