+++ tsconfig.json
@@ -1,0 +1,40 @@
+{
+  "include": ["./src"],
+  "exclude": ["node_modules", "dist"],
+  "compilerOptions": {
+    /* 基本选项 */
+    "target": "ES2017",                    // 指定 ECMAScript 目标版本: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', or 'ESNEXT'
+    "module": "commonjs",                  // 指定使用模块: 'commonjs', 'amd', 'system', 'umd' or 'es2015'
+    "lib": ["es2017"],                             // 指定要包含在编译中的库文件
+
+    "outDir": "./dist",                  // 指定输出目录
+    "rootDir": "./src",                       // 用来控制输出目录结构 --outDir.
+    "resolveJsonModule": true,
+    "noEmitHelpers": true,
+    "removeComments": true,                // 删除编译后的所有的注释
+    "importHelpers": true,                 // 从 tslib 导入辅助工具函数
+    /* 严格的类型检查选项 */
+    "strict": true,                        // 启用所有严格类型检查选项
+    "noImplicitAny": true,                 // 在表达式和声明上有隐含的 any类型时报错
+    "strictNullChecks": true,              // 启用严格的 null 检查
+    "noImplicitThis": true,                // 当 this 表达式值为 any 类型的时候，生成一个错误
+    "alwaysStrict": true,                  // 以严格模式检查每个模块，并在每个文件里加入 'use strict'
+
+    "noFallthroughCasesInSwitch": true,    // 报告 switch 语句的 fallthrough 错误。（即，不允许 switch 的 case 语句贯穿）
+
+    /* 模块解析选项 */
+    "moduleResolution": "node",            // 选择模块解析策略： 'node' (Node.js) or 'classic' (TypeScript pre-1.6)
+    "baseUrl": "./",                        // 用于解析非相对模块名称的基目录
+    "paths": {},                           // 模块名到基于 baseUrl 的路径映射的列表
+
+    "declaration": true,                   // 生成相应的 '.d.ts' 文件
+
+    /* Source Map Options */
+    "sourceMap": true,                     // 生成相应的 '.map' 文件
+    "sourceRoot": "",                    // 指定调试器应该找到 TypeScript 文件而不是源文件的位置
+    "mapRoot": "",                       // 指定调试器应该找到映射文件而不是生成文件的位置
+    /* 其他选项 */
+    "experimentalDecorators": true,        // 启用装饰器
+    "esModuleInterop": true,
+  }
+}
