--- typings/core/flag-pool.d.ts
@@ -1,14 +1,0 @@
-declare module '@flasco/cheat-core/src/core/flag-pool' {
-  import * as cv from 'opencv4nodejs';
-
-  class FlagPool {
-    private appMap: Object;
-    /**
-     * 获取处理成能被 opencv 使用的图片
-     * @param flagPath 文件地址
-     */
-    getFlag(flagPath: string): cv.Mat;
-  }
-
-  export = FlagPool;
-}
